Fixed spacing problem in inbox
There was a problem previously where the name of the replier would be cut off in the inbox. Changing the margin fixes that issue; there is however a tradeoff, in doing this there is a 5px loss per message in the chat because they use the same CSS class. Previously the margin was -10, but this gets cut off in the inbox, -5 is the lowest number that doesn't cut off the name but does have that trade-off. It's not all that noticeable though.