Fixed the text encroaching on the quoted message in inbox
There was a bug previously where the reply to a message would encroach on the original quoted message in the inbox. I fixed it by changing the margin around that quoted message by about 2px, and I changed the padding around that quoted message. This was important because it decreases the amount of space lost in chat. Unfortunately, because every message in both the inbox and chat uses the same CSS changing one affects the other, but by shrinking the padding I think the loss is limited to a few pixels per replied message in chat, and it's hardly noticeable.